Security e teste

+++ src/main/java/br/com/grupobrasileiro/sistema_grupo_brasileiro_backend/dto/user/form/EmployeeForm.java
@@ -59,6 +59,6 @@ public record EmployeeForm(
          * Este campo é obrigatório.
          */
         @NotNull(message = "The employee avatar cannot be null")
-        Integer avatar
+        Long avatar
 ) {
 }
